Made a working prototype
Made a working prototype with movement and blocks to break. It's very basic. I also started work on a damage layers system to help avoid friendlies damaging each other. It's a bit unnecessary at the moment but it will be a lot easier if I implement it early than if I make it later